// src/app/shared/models/product.ts
export class Product {
  $key: string;
  productId: number;
  productName: string;
productModel: string;
productColor:string;
productMaterial:string;
productMedidas:number;
productMedidas2:number;
productCategory: string;
productPrice: number;
productDescription: string;
productLocatar: string;
productState: string;
productDir: string;
productTel: string; 
  productImageUrl: string;
  productImageUrl2: string;
  productImageUrl3: string;
  productImageUrl4: string;
  productAdded: number;
  productQuatity: number;
  ratings: number;
  favourite: boolean;
  productSeller: string;
}